+++ fabric/main.py
@@ -18,11 +18,11 @@ import types
 
 from fabric import api, state  # For checking callables against the API, & easy mocking
 from fabric.contrib import console, files, project  # Ditto
-from fabric.network import denormalize, interpret_host_string, disconnect_all
-from fabric.state import commands, connections, env_options
+from fabric.network import denormalize, interpret_host_string, disconnect_all, normalize_to_string
+from fabric.state import commands, env_options
 from fabric.tasks import Task
 from fabric.utils import abort, indent
-
+from job_queue import JobQueue
 
 # One-time calculation of "all internal callables" to avoid doing this on every
 # check of a given fabfile callable (in is_classic_task()).
@@ -604,6 +604,7 @@ def get_hosts(command, cli_hosts, cli_roles, cli_exclude_hosts):
     # list if no hosts have been set anywhere.
     return _merge(state.env['hosts'], state.env['roles'], state.env['exclude_hosts'])
 
+
 def update_output_levels(show, hide):
     """
     Update state.output values as per given comma-separated list of key names.
@@ -621,12 +622,50 @@ def update_output_levels(show, hide):
             state.output[key] = False
 
 
+def requires_parallel(task):
+    """
+    Returns True if given ``task`` should be run in parallel mode.
+
+    Specifically:
+
+    * It's been explicitly marked with ``@parallel``, or:
+    * It's *not* been explicitly marked with ``@serial`` *and* the global
+      parallel option (``env.parallel``) is set to ``True``.
+    """
+    return (
+        (state.env.parallel and not getattr(task, 'serial', False))
+        or getattr(task, 'parallel', False)
+    )
+
+
 def _run_task(task, args, kwargs):
     # First, try class-based tasks
     if hasattr(task, 'run') and callable(task.run):
         return task.run(*args, **kwargs)
     # Fallback to callable behavior
     return task(*args, **kwargs)
+
+
+def _get_pool_size(task, hosts):
+    # Default parallel pool size (calculate per-task in case variables
+    # change)
+    default_pool_size = state.env.pool_size or len(hosts)
+    # Allow per-task override
+    pool_size = getattr(task, 'pool_size', default_pool_size)
+    # But ensure it's never larger than the number of hosts
+    pool_size = min((pool_size, len(hosts)))
+    # Inform user of final pool size for this task
+    if state.output.debug:
+        msg = "Parallel tasks now using pool size of %d"
+        print msg % state.env.pool_size
+    return pool_size
+
+
+def _parallel_tasks(commands_to_run):
+    return any(map(
+        lambda x: requires_parallel(crawl(x[0], state.commands)),
+        commands_to_run
+    ))
 
 
 def main():
@@ -751,6 +790,16 @@ Remember that -f can be used to specify fabfile path, and use -h for help.""")
             names = ", ".join(x[0] for x in commands_to_run)
             print("Commands to run: %s" % names)
 
+        # Import multiprocessing if needed, erroring out usefully if it can't.
+        if state.env.parallel or _parallel_tasks(commands_to_run):
+            try:
+                import multiprocessing
+            except ImportError, e:
+                msg = "At least one task needs to be run in parallel, but the\nmultiprocessing module cannot be imported:"
+                msg += "\n\n\t%s\n\n" % e
+                msg += "Please make sure the module is installed or that the above ImportError is\nfixed."
+                abort(msg)
+
         # At this point all commands must exist, so execute them in order.
         for name, args, kwargs, cli_hosts, cli_roles, cli_exclude_hosts in commands_to_run:
             # Get callable by itself
@@ -760,6 +809,14 @@ Remember that -f can be used to specify fabfile path, and use -h for help.""")
             # Set host list (also copy to env)
             state.env.all_hosts = hosts = get_hosts(
                 task, cli_hosts, cli_roles, cli_exclude_hosts)
+
+            # Get pool size for this task
+            pool_size = _get_pool_size(task, hosts)
+            # Set up job queue in case parallel is needed
+            jobs = JobQueue(pool_size)
+            if state.output.debug:
+                jobs._debug = True
+
             # If hosts found, execute the function on each host in turn
             for host in hosts:
                 # Preserve user
@@ -769,13 +826,43 @@ Remember that -f can be used to specify fabfile path, and use -h for help.""")
                 # Log to stdout
                 if state.output.running and not hasattr(task, 'return_value'):
                     print("[%s] Executing task '%s'" % (host, name))
-                # Actually run command
-                _run_task(task, args, kwargs)
+
+                # Handle parallel execution
+                if requires_parallel(task):
+                    # Grab appropriate callable (func or instance method)
+                    to_call = task
+                    if hasattr(task, 'run') and callable(task.run):
+                        to_call = task.run
+                    # Wrap in another callable that nukes the child's cached
+                    # connection object, if needed, to prevent shared-socket
+                    # problems.
+                    def inner(*args, **kwargs):
+                        key = normalize_to_string(state.env.host_string)
+                        state.connections.pop(key, "")
+                        to_call(*args, **kwargs)
+                    # Stuff into Process wrapper
+                    p = multiprocessing.Process(target=inner, args=args,
+                        kwargs=kwargs)
+                    # Name/id is host string
+                    p.name = state.env.host_string
+                    # Add to queue
+                    jobs.append(p)
+                # Handle serial execution
+                else:
+                    _run_task(task, args, kwargs)
+
                 # Put old user back
                 state.env.user = prev_user
+
+            # If running in parallel, block until job queue is emptied
+            if jobs:
+                jobs.close()
+                jobs.start()
+
             # If no hosts found, assume local-only and run once
             if not hosts:
                 _run_task(task, args, kwargs)
+
         # If we got here, no errors occurred, so print a final note.
         if state.output.status:
             print("\nDone.")