+++ docs/changelog.rst
@@ -25,6 +25,8 @@ would have also been included in the 1.2 line.
 Changelog
 =========
 
+* :feature:`19` Tasks may now be optionally executed in parallel. Please see
+  the :doc:`parallel execution docs </usage/parallel>` for details.
 * :bug:`182` During display of remote stdout/stderr, Fabric occasionally
   printed extraneous line prefixes (which in turn sometimes overwrote wrapped
   text.) This has been fixed.